enable cleandOldFiles to be scheduled
