+++ AccessibilityInsightsForAndroidService/app/src/main/java/com/microsoft/accessibilityinsightsforandroidservice/TempFileProvider.java
@@ -6,11 +6,9 @@ package com.microsoft.accessibilityinsightsforandroidservice;
 import android.content.Context;
 
 import androidx.annotation.NonNull;
-import androidx.work.Configuration;
-import androidx.work.ListenableWorker;
+import androidx.work.Data;
 import androidx.work.OneTimeWorkRequest;
 import androidx.work.WorkManager;
-import androidx.work.WorkRequest;
 import androidx.work.Worker;
 import androidx.work.WorkerParameters;
 
@@ -33,19 +31,21 @@ public class TempFileProvider {
   @NonNull public static final int tempFileLifetimeMillis = 5 * 60 * 1000; // 5 minutes
   @NonNull private File tempDir;
   @NonNull private Context context;
-  @NonNull private  WorkerManagerRunner workerManagerRunner;
 
-  public TempFileProvider(Context context, WorkerManagerRunner workerManagerRunner) {
+  public TempFileProvider(Context context) {
     this.context = context;
-    this.workerManagerRunner = workerManagerRunner;
     File cacheDir = context.getCacheDir();
     String tempDirPath = cacheDir.getAbsolutePath() + File.separator + tempDirName;
     this.tempDir = new File(tempDirPath);
   }
 
   public void cleanOldFilesBestEffort() {
+    cleanOldFilesBestEffort(tempDir);
+  }
+
+  private static void cleanOldFilesBestEffort(File tempDir) {
     long cutoffTime = new Date().getTime() - tempFileLifetimeMillis;
-    File[] files = this.tempDir.listFiles();
+    File[] files = tempDir.listFiles();
     if (files != null) {
       for (File file : files) {
         if (file.lastModified() < cutoffTime) {
@@ -66,25 +66,7 @@ public class TempFileProvider {
       writer.write(contents);
       writer.flush();
     }
-
-
-    //We need try to create our own configuration of the WorkManager
-    //Because it use to be created in the activity (but we dont have one)
-    //We are trying to move this part because now we are not abel to mock it (because its android level files)
-    /*Configuration myConfig = new Configuration.Builder()
-            .setMinimumLoggingLevel(android.util.Log.INFO)
-            .build();
-    WorkManager.initialize(context, myConfig);
-
-
-    OneTimeWorkRequest cleanFilesWorker = new OneTimeWorkRequest.Builder(CleanWorker.class)
-            .setInitialDelay(5, TimeUnit.MINUTES)
-            .build();
-    WorkManager.getInstance(context).enqueue(cleanFilesWorker);
-
-    WorkerManagerRunner workerManagerRunner = new WorkerManagerRunner(context);
-    workerManagerRunner.startWorkerManager().enqueue(workerManagerRunner.createTask());*/
-    workerManagerRunner.enqueueTask(CleanWorker.class);
+    scheduleCleanOldFiles(tempDir.getAbsolutePath());
     return tempFile;
   }
 
@@ -93,21 +75,28 @@ public class TempFileProvider {
     this.tempDir.mkdir();
   }
 
-  /*public class CleanWorker extends Worker {
+  private void scheduleCleanOldFiles(String tempDir){
+    Data inputData = new Data.Builder().putString("tempDir", tempDir).build();
+    OneTimeWorkRequest cleanFilesWorker = new OneTimeWorkRequest.Builder(CleanWorker.class)
+            .setInitialDelay(tempFileLifetimeMillis, TimeUnit.MILLISECONDS)
+            .setInputData(inputData)
+            .build();
+    WorkManager.getInstance(context).enqueue(cleanFilesWorker);
+  }
 
-    private final Context context;
+  public static class CleanWorker extends Worker {
+    private String tempDir;
 
-    public CleanWorker(@NonNull Context context, @NonNull WorkerParameters workerParams) {
+    public CleanWorker(@NonNull Context context, @NonNull  WorkerParameters workerParams) {
       super(context, workerParams);
-      this.context = context;
+      tempDir = workerParams.getInputData().getString("tempDir");
     }
 
     @NonNull
     @Override
-    public ListenableWorker.Result doWork() {
-      TempFileProvider tempFileProvider = new TempFileProvider(context.getCacheDir());
-      tempFileProvider.cleanOldFilesBestEffort();
-      return ListenableWorker.Result.success();
+    public Result doWork() {
+      cleanOldFilesBestEffort(new File(tempDir));
+      return Result.success();
     }
-  }*/
+  }
 }
